refactor(test): rebalance deployment_test.go (move init/HPA blocks to correct files)
The previous split left inferenceservice_deployment_test.go at 4717 lines
(76 percent of the total), which somewhat defeats the goal of better
navigation. This rebalance moves three Describe blocks that fit better
in other files.

Moves:

buildModelInitCommand: deployment_test.go -> storage_test.go
  Tests how the init container command is constructed for model storage,
  which is a storage concern. Belongs alongside buildCachedStorageConfig
  and buildEmptyDirStorageConfig.

shouldWarnMissingSkipModelInit: deployment_test.go -> storage_test.go
  Tests a warning about missing skip-model-init flag, which is again a
  model-init/storage concern, not deployment shape.

HPA Autoscaling: deployment_test.go -> scheduling_test.go
  HPA is autoscaling, which is scheduling-adjacent. The scheduling file
  was only 100 lines (resolvePriorityClassName, resolveEffectivePriority);
  adding HPA balances it and groups all scheduling/scaling concerns together.

Resulting file sizes:

inferenceservice_deployment_test.go    4044 lines  (was 4717)
inferenceservice_reconcile_test.go      810 lines  (unchanged)
inferenceservice_scheduling_test.go     687 lines  (was 100)
inferenceservice_service_test.go        281 lines  (unchanged)
inferenceservice_storage_test.go        388 lines  (was 302)
